Refactor AddNewTask with new Task id generator

+++ src/org/ecs160/a2/utils/Database.java
@@ -12,10 +12,25 @@ import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 public class Database {
+
     static Storage db = Storage.getInstance();
 
     public static void init() {
         Util.register(Task.OBJECT_ID, Task.class);
+        Database.test();
+    }
+
+    public static int generateID(String key) {
+        int id = readID(key);
+        Log.p(String.valueOf(id));
+        db.writeObject(key, id + 1);
+        return id;
+    }
+
+    public static int readID(String key) {
+        Integer id = (Integer) db.readObject(key);
+        if (id == null) return 0;
+        return id;
     }
 
     public static void write(String key, Object val) {
@@ -41,11 +56,7 @@ public class Database {
     }
 
     public static void update(String key, Object val) {
-        switch (key) {
-            case Task.OBJECT_ID:
-                updateTask(key, (Task) val);
-                break;
-        }
+        if (Task.OBJECT_ID.equals(key)) updateTask(key, (Task) val);
     }
 
     public static Vector<Object> readAll(String key) {
@@ -55,11 +66,7 @@ public class Database {
     }
 
     public static void delete(String key, String id) {
-        switch (key) {
-            case Task.OBJECT_ID:
-                deleteTask(key, id);
-                break;
-        }
+        if (Task.OBJECT_ID.equals(key)) deleteTask(key, id);
     }
 
     public static void deleteTask(String key, String title) {
@@ -74,24 +81,29 @@ public class Database {
     }
 
     public static void test() {
-        Task t1 = new Task("test 1", "yee");
-        t1.start();
-        Task t2 = new Task("test 2", "yoo");
-        Task t3 = new Task("test 3", "yaa");
+//        Task t1 = new Task("test 1", "yee");
+//        t1.start();
+//        Task t2 = new Task("test 2", "yoo");
+//        Task t3 = new Task("test 3", "yaa");
         String key = Task.OBJECT_ID;
 
         List<Task> tests = new ArrayList<>();
-        tests.add(t1);
-        tests.add(t2);
+//        tests.add(t1);
+//        tests.add(t2);
 
-        t1.stop();
+//        if (tests != null) {
+//            for (Task t : tests) Log.p(t.getTitle() + t.getID());
+//        }
+
+//        t1.stop();
 //        deleteAll(key);
 //        Database.writeAll(key, (List) tests);
 //        Database.write(key, t3);
         List<Task> vec = (List) readAll(key);
 
         if (vec != null) {
-            for (Task t : vec) Log.p(t.getTitle());
+            for (Task t : vec)
+                Log.p(t.getTitle() + " " + t.getID());
         }
     }
